+++ dns_zone.php
@@ -19,27 +19,35 @@
 		
 		//TODO Ressource record list of zone
 	} elseif($_GET['section'] == 'insert_add') {
-		$dns_zone = new DnsZone(false, (int)$_SESSION['user_id'], $_POST['name'], $_POST['pri_dns'], $_POST['sec_dns'],
-								(int)$_POST['serial'], (int)$_POST['refresh'], (int)$_POST['retry'],
-								(int)$_POST['expire'], (int)$_POST['ttl']);
-		if($dns_zone->store()) {
-			$message[] = array('Neue DNS-Zone '.$_POST['name'].' wurde eingetragen.', 1);
+		if(Permission::checkPermission(PERM_USER)) {
+			$dns_zone = new DnsZone(false, (int)$_SESSION['user_id'], $_POST['name'], $_POST['pri_dns'], $_POST['sec_dns'],
+									(int)$_POST['serial'], (int)$_POST['refresh'], (int)$_POST['retry'],
+									(int)$_POST['expire'], (int)$_POST['ttl']);
+			if($dns_zone->store()) {
+				$message[] = array('Neue DNS-Zone '.$_POST['name'].' wurde eingetragen.', 1);
+			} else {
+				$message[] = array('Neue DNS-Zone '.$_POST['name'].' konnte nicht eingetragen werden.', 2);
+			}
+			Message::setMessage($message);
+			
+			header('Location: ./dns_zone.php');
 		} else {
-			$message[] = array('Neue DNS-Zone '.$_POST['name'].' konnte nicht eingetragen werden.', 2);
+			Permission::denyAccess(PERM_USER);
 		}
-		Message::setMessage($message);
-		
-		header('Location: ./dns_zone.php');
 	} elseif($_GET['section'] == 'delete') {
 		$dns_zone = new DnsZone((int)$_GET['dns_zone_id']);
 		$dns_zone->fetch();
-		$dns_zone_name = $dns_zone->getName();
-		$dns_zone->delete();
-		
-		$message[] = array('Die DNS-Zone '.$dns_zone_name.' wurde gelöscht.', 1);
-		Message::setMessage($message);
-		
-		header('Location: ./dns_zone.php');
+		if(permission::checkIfUserIsOwnerOrPermitted(PERM_ROOT, $dns_zone->getUserId())) {
+			$dns_zone_name = $dns_zone->getName();
+			$dns_zone->delete();
+			
+			$message[] = array('Die DNS-Zone '.$dns_zone_name.' wurde gelöscht.', 1);
+			Message::setMessage($message);
+			
+			header('Location: ./dns_zone.php');
+		} else {
+			Permission::denyAccess(PERM_ROOT, $dns_zone->getUserId());
+		}
 	} else {
 		$dns_zone_list = new DnsZoneList();
 		$smarty->assign('dns_zone_list', $dns_zone_list->getDnsZoneList());
